Create LICENSE
This license states that you may use this code for whatever you want as long as you credit me as the developer. Meaning you may NOT distribute this code to another user claiming it to be yours. This includes making money off this code. If you are selling this without giving me FULL credit as the developer, legal action can and will be taken against you.

Using my tutorials for help is fine, but you cannot straight copy my code and claim it to be yours.

Please make sure you understand this license before continuing.